Refactor PlaintextField out to common component, Refactor Provenance to only take in required information
