+++ app/src/components/ArtworkInfo.tsx
@@ -142,8 +142,7 @@ class ArtworkInfo extends React.Component<ArtworkInfoProps, ArtworkInfoState> {
             <div className="text-center">
               <ButtonGroup>
                 <Documents documents={fields.documents}/>
-                <Provenance registry={this.props.contracts.ArtifactRegistry} metaUri={this.props.artwork.metaUri}
-                  tokenId={this.props.id} contracts={this.props.contracts} accounts={this.props.accounts}/>
+                <Provenance registry={this.props.contracts.ArtifactRegistry} tokenId={this.props.id}/>
               </ButtonGroup>
             </div>
